pyproject.toml
[project]
name = "agents-tutorial"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.0.20",
    "deepagents>=0.5.2",
    "flashrank>=0.2.10",
    "gradio>=6.12.0",
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.6",
    "langchain>=1.2.15",
    "langchain-chroma>=1.0.0",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.12",
    "langgraph",
    "langgraph-swarm>=0.1.0",
    "mcp-neo4j-cypher>=0.5.3",
    "neo4j>=6.1.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "tqdm>=4.67.3",
]
